feat: create @cache decorator to manage Redis cache via UseCases container
