+++ backend/bloom/infra/repositories/repository_excursion.py
@@ -37,7 +37,7 @@ class ExcursionRepository:
     def get_excursions_by_vessel_id(self, session: Session, vessel_id: int) -> List[Excursion]:
         """Recheche l'excursion en cours d'un bateau, c'est-à-dire l'excursion qui n'a pas de date d'arrivée"""
         stmt = select(sql_model.Excursion).where(sql_model.Excursion.vessel_id == vessel_id)
-        result = session.execute(stmt).scalars()
+        result = session.execute(stmt).scalars().all()
         if not result:
             return []
         return [ExcursionRepository.map_to_domain(r) for r in result]
@@ -46,7 +46,7 @@ class ExcursionRepository:
         """Recheche l'excursion en cours d'un bateau, c'est-à-dire l'excursion qui n'a pas de date d'arrivée"""
         stmt = select(sql_model.Excursion).where((sql_model.Excursion.vessel_id == vessel_id)
                                                  & (sql_model.Excursion.id == excursion_id))
-        result = session.execute(stmt).scalar()
+        result = session.execute(stmt).scalar_one_or_none()
         if not result:
             return None
         return ExcursionRepository.map_to_domain(result)
